refactor: move isMedia to utils
